various changes and fixes
-dscfsprite no longer uses scaleX and scaleY, but instead scales using new "size" parameter
-width and height are now floats (as they ought to be)
-added open and save file to utils
-added "onAfterDepthSortedRender" to window (and arguably made the transformgizmo renderer worse)
-added "open file" button for any string parameter

+++ GXPEngine/GXPEngine/AnimationSprite.cs
@@ -108,7 +108,7 @@ namespace GXPEngine
 		/// <summary>
 		/// Gets or sets the sprite's width in pixels.
 		/// </summary>
-		override public int width {
+		override public float width {
 			get { 
 				if (_texture != null) return (int)Math.Abs(_texture.width * _scaleX * _frameWidth);
 				return 0;
@@ -124,7 +124,7 @@ namespace GXPEngine
 		/// <summary>
 		/// Gets or sets the sprite's height in pixels.
 		/// </summary>
-		override public int height {
+		override public float height {
 			get { 
 				if (_texture != null) return (int)Math.Abs(_texture.height * _scaleY * _frameHeight);
 				return 0;
