+++ GXPEngine/GXPEngine/UI/SliderPanel.cs
@@ -50,7 +50,7 @@ namespace GXPEngine.UI
             var oldRange = game.RenderRange;
             var c = GetArea();
             Vector3 truepos = TransformPoint(0, 0, 0);
-            game.SetViewport((int)(truepos.x-c[0]), (int)(truepos.y-c[3]), width, height, false);
+            game.SetViewport((int)(truepos.x-c[0]), (int)(truepos.y-c[3]), (int)width, (int)height, false);
             base.Render(glContext);
             game.SetViewport((int)oldRange.left, (int)oldRange.top, (int)oldRange.width, (int)oldRange.height);
         }
